Added cache manager [ehcache] and added more options to the config file.
Cached user information is purged periodically to allow changes to
/etc/password or /etc/shadow to be detected and followed by NFSRODS. The
following config option controls the refresh time.

   - user_information_refresh_time_in_minutes

Stat information obtained through a network call to iRODS is now cached
for a very short period to increase the speed of NFSRODS and make it
useable. The following config option controls the refresh time.

   - file_information_refresh_time_in_seconds